fix(daily-pageviews): correctly truncate days at midnight, not at current time
when calling the daily pageviews service in the middle of the traffic peak (i.e. at US west coast lunch time), the numbers for the last day in the reporting period were suspicously low. This was due to an incorrect conversion between the current timestamp and the cutoff time for the end of the day

see https://cq-dev.slack.com/archives/C01F5TQV0SF/p1635756602194500?thread_ts=1635283992.140100&cid=C01F5TQV0SF