fix(checkpoints): calculate percentages based on weight and not just on raw event count
Fixes two issues with the calculation of percentages: (1) percentages were based on event counts, disregarding weight. This lead to an overcounting of `unsupported` and `noscript` checkpoints which always come at weight `1`. (2) the total for `any` overestimated the total number of events by about 100%, leading to a huge dropoff for `top`. This has been fixed by counting each `id` only once when estimating the number of events of any kind.