feat(bundles): enable pagination
pagination uses the `after` and `limit` URL parameters and prevents response sizes from
getting too big. Run the query a first time without the `after` URL parameter and it will
return the first `limit` rows. After that, take the value of `id` of the last record received
and use it as the value of the `after` URL parameter. Repeat the process until no more rows
are returned